+++ packages/input-number/src/input-number.vue
@@ -7,7 +7,6 @@
   >
     <el-input
       v-model="currentValue"
-      @onchange="handleChange"
       :disabled="disabled"
       :size="size"
       :number="true"
@@ -41,7 +40,7 @@
     name: 'ElInputNumber',
     props: {
       value: {
-        type: Number
+        default: 1
       },
       step: {
         type: Number,
@@ -91,42 +90,86 @@
     },
     data() {
       return {
-        currentValue: null,
+        currentValue: this.value,
         inputActive: false
       };
     },
     watch: {
-      value: {
-        immediate: true,
-        handler(val) {
-          this.currentValue = val;
-        }
-      },
-      currentValue(val) {
-        if (!isNaN(parseInt(val, 10))) {
-          this.$emit('input', parseInt(val, 10));
+      currentValue(newVal, oldVal) {
+        if (!isNaN(newVal) && newVal <= this.max && newVal >= this.min) {
+          this.$emit('change', newVal);
+          this.$emit('input', newVal);
+        } else {
+          this.$nextTick(() => {
+            this.currentValue = oldVal;
+          });
         }
       }
     },
     computed: {
       minDisabled() {
-        return this.value - this.step < this.min;
+        return this.currentValue - this.step < this.min;
       },
       maxDisabled() {
-        return this.value + this.step > this.max;
+        return this.currentValue + this.step > this.max;
       }
     },
     methods: {
+      accSub(arg1, arg2) {
+        var r1, r2, m, n;
+        try {
+          r1 = arg1.toString().split('.')[1].length;
+        } catch (e) {
+          r1 = 0;
+        }
+        try {
+          r2 = arg2.toString().split('.')[1].length;
+        } catch (e) {
+          r2 = 0;
+        }
+        m = Math.pow(10, Math.max(r1, r2));
+        n = (r1 >= r2) ? r1 : r2;
+        return parseFloat(((arg1 * m - arg2 * m) / m).toFixed(n));
+      },
+      accAdd(arg1, arg2) {
+        var r1, r2, m, c;
+        try {
+          r1 = arg1.toString().split('.')[1].length;
+        } catch (e) {
+          r1 = 0;
+        }
+        try {
+          r2 = arg2.toString().split('.')[1].length;
+        } catch (e) {
+          r2 = 0;
+        }
+        c = Math.abs(r1 - r2);
+        m = Math.pow(10, Math.max(r1, r2));
+        if (c > 0) {
+          var cm = Math.pow(10, c);
+          if (r1 > r2) {
+            arg1 = Number(arg1.toString().replace('.', ''));
+            arg2 = Number(arg2.toString().replace('.', '')) * cm;
+          } else {
+            arg1 = Number(arg1.toString().replace('.', '')) * cm;
+            arg2 = Number(arg2.toString().replace('.', ''));
+          }
+        } else {
+          arg1 = Number(arg1.toString().replace('.', ''));
+          arg2 = Number(arg2.toString().replace('.', ''));
+        }
+        return (arg1 + arg2) / m;
+      },
       increase() {
-        if (this.value + this.step > this.max || this.disabled) return;
-        this.currentValue += this.step;
+        if (this.currentValue + this.step > this.max || this.disabled) return;
+        this.currentValue = this.accAdd(this.step, this.currentValue);
         if (this.maxDisabled) {
           this.inputActive = false;
         }
       },
       decrease() {
-        if (this.value - this.step < this.min || this.disabled) return;
-        this.currentValue -= this.step;
+        if (this.currentValue - this.step < this.min || this.disabled) return;
+        this.currentValue = this.accSub(this.currentValue, this.step);
         if (this.minDisabled) {
           this.inputActive = false;
         }
@@ -140,9 +183,6 @@
         if (!this.disabled && !disabled) {
           this.inputActive = false;
         }
-      },
-      handleChange(value) {
-        this.$emit('onchange', value);
       }
     }
   };